LiveBackend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
bcrypt
openai
mysql.connector
werkzeug
MIMEMultipart
MIMEText
load_dotenv
base64
hashlib
os
uuid
datetime
typing
hashlib
cryptography